Validation
Added php validation to the add listing for. Moved the image upload validation out of the process-add-listing script, into the validation script. Rewrote the image upload script so that images are store in a temp folder on the server before processing by the process-add-listing-script, this tidied things up a lot.